refactor: streamline Makefile and add diff functionality with tests
