+++ main.go
@@ -1,7 +1,6 @@
 package main
 
 import (
-	"bufio"
 	"bytes"
 	"crypto/sha256"
 	"fmt"
@@ -12,7 +11,6 @@ import (
 	"runtime"
 	"strings"
 
-	"github.com/cubicdaiya/gonp"
 	"github.com/jessevdk/go-flags"
 	"github.com/mackerelio/checkers"
 	"github.com/mackerelio/golib/pluginutil"
@@ -20,6 +18,13 @@ import (
 
 var version string
 var commit string
+
+const (
+	OK = iota
+	WARNING
+	CRITICAL
+	UNKNOWN
+)
 
 type Opt struct {
 	Args       []string
@@ -40,7 +45,7 @@ func (opt *Opt) cmd(file *os.File) error {
 	}
 	err := cmd.Wait()
 	if err != nil {
-		return fmt.Errorf("%s - %s", err, stderr.String())
+		return fmt.Errorf("%w - %s", err, stderr.String())
 	}
 	return nil
 }
@@ -69,7 +74,7 @@ func (opt *Opt) run() *checkers.Checker {
 	err = opt.cmd(newFile)
 	if err != nil {
 		newFile.Close()
-		os.Remove(newFile.Name())
+		_ = os.Remove(newFile.Name())
 		return checkers.Critical(err.Error())
 	}
 
@@ -114,77 +119,6 @@ func (opt *Opt) run() *checkers.Checker {
 	return checkers.Critical(diffMsg)
 }
 
-func buildNoDifferenceMsg(filename string) (string, error) {
-	file, err := openRD(filename)
-	if err != nil {
-		return "", err
-	}
-	defer file.Close()
-
-	fileinfo, err := file.Stat()
-	if err != nil {
-		return "", err
-	}
-	b := make([]byte, 128)
-	count, err := file.Read(b)
-	if err != nil {
-		return "", err
-	}
-	o := string(strings.TrimRight(string(b[0:count]), "\r\n"))
-	if fileinfo.Size() > 128 {
-		return fmt.Sprintf("no difference: ```%s...```", o), nil
-	}
-	return fmt.Sprintf("no difference: ```%s```", o), nil
-}
-
-func buildDiffMsg(diff string) string {
-	o := diff
-	if len(diff) > 512 {
-		o = diff[0:512]
-	}
-	o = strings.TrimRight(o, "\r\n")
-
-	if len(diff) > 512 {
-		return fmt.Sprintf("found difference: ```%s...```", o)
-	}
-	return fmt.Sprintf("found difference: ```%s```", o)
-}
-
-func getLines(filename string) ([]string, error) {
-	file, err := openRD(filename)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	scanner.Buffer(make([]byte, 64*1024), 10*1024*1024)
-	lines := make([]string, 0)
-	for scanner.Scan() {
-		lines = append(lines, scanner.Text())
-	}
-	if err := scanner.Err(); err != nil {
-		return nil, err
-	}
-	return lines, nil
-}
-
-func diff(prev, new string) (string, error) {
-	prevLines, err := getLines(prev)
-	if err != nil {
-		return "", err
-	}
-	newLines, err := getLines(new)
-	if err != nil {
-		return "", err
-	}
-
-	diff := gonp.New(prevLines, newLines)
-	diff.Compose()
-
-	return diff.SprintUniHunks(diff.UnifiedHunks()), nil
-}
-
 func fileExists(filename string) bool {
 	_, err := os.Stat(filename)
 	return err == nil
@@ -207,16 +141,19 @@ func main() {
 			runtime.GOARCH,
 			runtime.Version(),
 			commit)
-		os.Exit(0)
-	}
-	if err != nil {
-		fmt.Fprintf(os.Stderr, "%v\n", err)
-		os.Exit(1)
-	}
-	if len(args) == 0 {
+		os.Exit(OK)
+	} else if flags.WroteHelp(err) {
+		fmt.Fprintf(os.Stdout, "%v\n", err)
+		os.Exit(OK)
+	} else if len(args) == 0 {
+		fmt.Fprintf(os.Stderr, "command is required\n")
 		psr.WriteHelp(os.Stderr)
-		os.Exit(1)
+		os.Exit(UNKNOWN)
+	} else if err != nil {
+		fmt.Fprintf(os.Stderr, "%v\n", err)
+		os.Exit(UNKNOWN)
 	}
+
 	opt.Args = []string{}
 	opt.Command = args[0]
 	if len(args) > 1 {
