+++ main_test.go
@@ -35,7 +35,10 @@ func TestCmd(t *testing.T) {
 		t.Fatalf("cmd failed: %v", err)
 	}
 
-	file.Seek(0, 0) // Reset file pointer to the beginning
+	_, err := file.Seek(0, 0) // Reset file pointer to the beginning
+	if err != nil {
+		t.Fatalf("Failed to seek file: %v", err)
+	}
 	scanner := bufio.NewScanner(file)
 	if scanner.Scan() {
 		output := scanner.Text()
@@ -77,89 +80,7 @@ func createTempFile(t *testing.T, content string) (*os.File, func()) {
 		t.Fatalf("Failed to close temp file: %v", err)
 	}
 	return tmpfile, func() {
-		os.Remove(tmpfile.Name())
-	}
-}
-
-func TestBuildNoDifferenceMsg(t *testing.T) {
-	shortMsg := "This is a test message."
-	file, cleanup := createTempFile(t, shortMsg+"\n")
-	defer cleanup()
-	expected := "no difference: ```" + shortMsg + "```"
-	result, err := buildNoDifferenceMsg(file.Name())
-	if err != nil {
-		t.Fatalf("buildNoDifferenceMsg failed: %v", err)
-	}
-	if result != expected {
-		t.Errorf("Expected %q, got %q", expected, result)
-	}
-}
-
-func TestBuildNoDifferenceMsgLargeResult(t *testing.T) {
-	largeMsg := strings.Repeat("A", 600)
-	file, cleanup := createTempFile(t, largeMsg+"\n")
-	defer cleanup()
-
-	expected := "no difference: ```" + largeMsg[:128] + "...```"
-	result, err := buildNoDifferenceMsg(file.Name())
-	if err != nil {
-		t.Fatalf("buildNoDifferenceMsg failed: %v", err)
-	}
-	if result != expected {
-		t.Errorf("Expected %q, got %q", expected, result)
-	}
-}
-
-func TestBuildDiffMsg(t *testing.T) {
-	diff := "This is a test diff message.\nThis is the second line of the diff.\n"
-	expected := "found difference: ```This is a test diff message.\nThis is the second line of the diff.```"
-	result := buildDiffMsg(diff)
-	if result != expected {
-		t.Errorf("Expected %q, got %q", expected, result)
-	}
-}
-
-func TestBuildDiffMsgLargeResult(t *testing.T) {
-	diff := strings.Repeat("A", 600)
-	expected := "found difference: ```" + diff[:512] + "...```"
-	result := buildDiffMsg(diff)
-	if result != expected {
-		t.Errorf("Expected %q, got %q", expected, result)
-	}
-}
-
-func TestDiff(t *testing.T) {
-	file1, cleanup1 := createTempFile(t, "Hello, World!\n")
-	defer cleanup1()
-
-	file2, cleanup2 := createTempFile(t, "Hello, Go!\n")
-	defer cleanup2()
-
-	diffResult, err := diff(file1.Name(), file2.Name())
-	if err != nil {
-		t.Fatalf("diff failed: %v", err)
-	}
-	if diffResult == "" {
-		t.Errorf("Expected a diff result, got empty string")
-	}
-	if !strings.Contains(diffResult, "Hello, World!") || !strings.Contains(diffResult, "Hello, Go!") {
-		t.Errorf("Diff result does not contain expected content: %q", diffResult)
-	}
-}
-
-func TestDiffNoDifference(t *testing.T) {
-	file1, cleanup1 := createTempFile(t, "Hello, World!\n")
-	defer cleanup1()
-
-	file2, cleanup2 := createTempFile(t, "Hello, World!\n")
-	defer cleanup2()
-
-	diffResult, err := diff(file1.Name(), file2.Name())
-	if err != nil {
-		t.Fatalf("diff failed: %v", err)
-	}
-	if diffResult != "" {
-		t.Errorf("Expected no diff result, got: %q", diffResult)
+		_ = os.Remove(tmpfile.Name())
 	}
 }
